Fix label id conflict error in LIPO mode with very large modules
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google/gcc-4_7@192639 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/dwarf2out.h
@@ -89,8 +89,8 @@ typedef struct GTY(()) dw_fde_struct {
   int dw_fde_switch_cfi_index; /* Last CFI before switching sections.  */
   HOST_WIDE_INT stack_realignment;
 
-  unsigned funcdef_number;
-  unsigned fde_index;
+  unsigned long funcdef_number;
+  unsigned long fde_index;
 
   /* Dynamic realign argument pointer register.  */
   unsigned int drap_reg;
